refactor: split manager media ingest router into read and write modules
